refactor: Rename and relocate app.tsx to app.js
