+++ __tests__/App.test.js
@@ -4,7 +4,7 @@
 
 import 'react-native';
 import React from 'react';
-import App from '../App';
+import App from '../src/App';
 
 // Note: import explicitly to use the types shiped with jest.
 import {it} from '@jest/globals';
